refactor(indx): #33 Remove unused .indx
The .indx attribute was implemented to keep compatibility with metaclass, and assumes the is a "NAME" column in the dataframe.

However since the defacto way of loading is to use read(path, index="NAME") and metaclass isn't really used any more, there is (and never was?) a use for .indx

This commit removes the attribute for good.