Dont log errors on read (#99)
* Change logger level to debug when raising an exception in read_tfs
From @st-walker :
An exception is not really an error, at least not in Python, so when I am
explicitly relying on for example the raising of a `TfsFormatError` in
`read_tfs` (before trying a different loader for what is presumably a different
file type), seeing:

        LOGGER.error(f"No column names in file {tfs_file_path.absolute()}, aborting")

is not really welcome or appropriate...  it's not even immediately clear to the
user where this talk of "aborting" is coming from due to the way the logger
formatter is configured, so it is quite alarming and annoying to be doing very
normal code and seeing "aborting"...  This should at most be `LOGGER.debug`, not
`LOGGER.debug`, I would suggest.

* Update reader.py

removed logging completely. Can be done by whoever excepts the error. All info is in it's message.

* Update frame.py

Switched from debug to error logging here as well.

* Update test_reader.py

fixed test to not check caplog

* Update test_writer.py

fixed test to check for DEBUG messages and nor ERROR.

* remove deprecated numpy.str in favor of built-in str

* up version number for new release once merged

Co-authored-by: stwalker <stuart.walker@desy.de>
Co-authored-by: Felix Soubelet <felix.soubelet@protonmail.com>